feat(tracing): add comprehensive tests for OpenTelemetry span starters and attributes
